Fix for skipping null case

+++ Scripts/delete_links.py
@@ -13,6 +13,8 @@ if __name__ == "__main__":
         index_links = json.load(fp)
 
         for link_path in index_links.keys():
+            if link_path is None or link_path is "null":
+                continue  # skip null case
             print(f"Removing {link_path}")
             os.unlink(link_path)
     os.remove("../index_links.json")
